Implement instrument detail page
Shows list of all in-stock instruments of that type.
Tweaks to Schema refs (from string to Schema object) to fix "Schema not
registered" bug - unsure why this fixes it. Bug only occured when deep-populating
the modelinstance.model.manufacturer field (but not the
modelinstance.model.instrument field for some reason).
Tweaked Error interfact extension to be global.